+++ test/Stormpath.Configuration.Test/ModifiedConfigTestCases/YamlTestCase.cs
@@ -97,6 +97,7 @@ web:
     uri: ""/register1""
     nextUri: ""/1""
     autoLogin: true
+    emailVerificationRequired: true
     form:
       fields:
         email: