+++ changelog.md
@@ -1,11 +1,13 @@
 # Changelog
 
-## Version 7.0.0
+### Version 7.0.0
 
 This version supports the migration of ASP.NET applications from Stormpath to Okta. For more information, see the master changelog (todo).
 
-## Breaking changes
+#### Breaking changes
 
+* The `web.changePassword.enabled` and `web.verifyEmail.enabled` settings are now **false** by default.
+* The `web.register.emailVerificationRequired` setting was added.
 * All configuration properties start with `okta.*` instead of `stormpath.*`. For most applications, this won't cause any issues.
 * JSON/YAML configuration can now be loaded from `okta.json` or `okta.yaml` **or** `stormpath.json`/`stormpath.yaml`. For backwards compatibility, either filename will work. (Note: the home directory search location is still `~/.stormpath`, for maximum backwards compatibility.)
 * Environment variable configuration can now be specified using `OKTA_` environment variables in addition to `STORMPATH_` environment variables. For backwards compatibility, either string format will work.